// src/models/generic/item.ts
export interface ItemModel {
  recid?: number;
  ptycde?: string;
  olditmcde?: string;
  saldisact?: string;
  purdisact?: string;
  field01?: string;
  field02?: string;
  field03?: string;
  field04?: string;
  field05?: string;
  field06?: string;
  field07?: string;
  field08?: string;
  field09?: string;
  field10?: string;
  field11?: string;
  field12?: string;
  field13?: string;
  field14?: string;
  field15?: string;
  field16?: string;
  field17?: string;
  field18?: string;
  field19?: string;
  field20?: string;
  scpwddis?: number;
  supdsc?: string;
  itmcde?: string;
  itmdsc?: string;
  untmea?: string;
  untmea2?: string;
  conver?: number;
  untcst1?: number;
  untcst2?: number;
  crilvl?: number;
  remarks?: string;
  wardsc?: string;
  avecst?: number;
  maxlvl?: number;
  status?: string;
  lstcst?: number;
  untcst?: number;
  untprc?: number;
  brndsc?: string;
  itmcladsc?: string;
  itmmdl?: string;
  supcde?: string;
  warcde?: string;
  brncde?: string;
  itmclacde?: string;
  salum?: string;
  srtum?: string;
  recum?: string;
  prtum?: string;
  invum?: string;
  barcde?: string;
  cstdebcde?: string;
  cstcrecde?: string;
  inactive?: number;
  itmtyp?: string;
  cgsactcde?: string;
  salactcde?: string;
  invactcde?: string;
  srtactcde?: string;
  multium?: number;
  reqsernum?: number;
  taxcde?: string;
  prtactcde?: string;
  puractcde?: string;
  purtaxcde?: string;
  salewtcde?: string;
  purewtcde?: string;
  salevatcde?: string;
  purevatcde?: string;
  salcur?: string;
  purcur?: string;
  itmrem1?: string;
  itmrem2?: string;
  itmrem3?: string;
  itmbal?: number;
  strqty?: string;
  chknontrd?: number;
  package?: number;
  rebdte?: string;
  itmprt?: string;
  gldepcde?: string;
  reqbatchnum?: number;
  logdte?: string;
  itmsubclacde?: string;
  linenum?: number;
  saltarget01?: number;
  saltarget02?: number;
  saltarget03?: number;
  saltarget04?: number;
  saltarget05?: number;
  saltarget06?: number;
  saltarget07?: number;
  saltarget08?: number;
  saltarget09?: number;
  saltarget10?: number;
  saltarget11?: number;
  saltarget12?: number;
  recumcon?: string;
  itmdeptcde?: string;
  itmmerchcatcde?: string;
  itmhierarchy?: string;
  itmclass?: string;
  itmtag?: string;
  itmsrc?: string;
  itmpalletcap?: number;
}