// src/models/generic/bin.ts
export interface BinModel {
  recid?: number;
  binnum?: string;
  batchnum?: string;
  itmcde?: string;
  warcde?: string;
  mfgdte?: string;
  expdte?: string;
  createddte?: string;
  inactivedte?: string;
  inactive?: string;
  hldoutgoing?: number;
  hldtypcde?: string;
  hldtypdsc?: string;
  warcdenum?: string;
  warloccde?: string;
  binrcklvl?: string;
  binrckzone?: string;
  strglvlcde?: string;
  strglvldsc?: string;
  brncde?: string;
  brndsc?: string;
  itmclacde?: string;
  itmcladsc?: string;
}